whatrag/pyproject.toml
[project]
name = "whatrag"
version = "0.1.0"
description = "WhatsApp chat RAG and persona simulation using LLMs"
readme = "README.md"
requires-python = "~=3.10.0"
dependencies = [
    "faiss-cpu>=1.13.2",
    "faiss-gpu>=1.7.2; sys_platform == 'linux'",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "python-dotenv>=1.2.1",
    "mcp[cli]>=1.26.0",
]
